Phase-4/backend/start_server.py
import os
from dotenv import load_dotenv
import uvicorn

# Load environment variables from .env file
load_dotenv('../.env')

if __name__ == "__main__":
    uvicorn.run(
        "main:app",
        host="0.0.0.0",
        port=8000,
        reload=True
    )