Phase-4/backend/pyproject.toml
[tool.poetry]
name = "todo-backend"
version = "0.1.0"
description = "Backend API for the Todo Web Application"
authors = ["Developer <developer@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
sqlmodel = "^0.0.8"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
alembic = "^1.13.1"
sqlalchemy = "^2.0.23"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.0"
httpx = "^0.25.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.11.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.uvicorn]
host = "0.0.0.0"
port = 8000
reload = true
factory = true
app_dir = "."

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"